third_party/WebKit/LayoutTests/platform/win/fast/forms/select/popup-menu-key-operations-expected.txt

PASS picker._selectElement.value is "1"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "4"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "4"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "1"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "0"
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "0"
PASS menuElement.value is "bar"
PASS window.internals.pagePopupWindow is null
PASS menuElement.value is "foo"
PASS picker._selectElement.value is "0"
PASS picker._selectElement.value is "1"
PASS menuElement.value is "foo"
PASS window.internals.pagePopupWindow is null
PASS menuElement.value is "bar"
PASS picker._selectElement.value is "1"
PASS picker._selectElement.value is "4"
PASS menuElement.value is "bar"
PASS window.internals.pagePopupWindow is null
PASS menuElement.value is "garply"
PASS picker._selectElement.value is "4"
PASS picker._selectElement.value is "0"
PASS menuElement.value is "garply"
PASS window.internals.pagePopupWindow is null
PASS menuElement.value is "garply"
PASS picker._selectElement.value is "4"
PASS picker._selectElement.value is "0"
PASS menuElement.value is "garply"
PASS picker._selectElement.value is "0"
PASS menuElement.value is "garply"
PASS window.internals.pagePopupWindow is null
PASS menuElement.value is "foo"
PASS successfullyParsed is true

TEST COMPLETE
